Stage setup and thread
I will be using a constantly running thread that will refresh the window 60. This makes it much easier to update the screen as the AI moves through the maze and also makes the programming of the AI itself easier as it can look at each scenario in a new way and decide the best move from that point. This is the way AStar and most other popular AIs work