Fixed too many explorers bug
the getVertices() function was returning multiple points along the same line. as a result, the children of these explorers would double back on each other as they were technically not going back on themselves, but simply being sent along the same route as another explorer. To resolve this, I have adapted the function to only return the "interesting tile" (either a corner, dead end or junction) that it comes across in each direction.